refactor: Add spi support for KindLoader
