+++ annotations/kind-annotations/src/main/java/io/dekorate/kind/KindImageLoader.java
@@ -1,0 +1,35 @@
+/**
+ * Copyright 2018 The original authors.
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * 
+**/
+
+package io.dekorate.kind;
+
+import io.dekorate.ImageLoader;
+import io.dekorate.Logger;
+import io.dekorate.LoggerFactory;
+import io.dekorate.project.Project;
+import io.dekorate.utils.Exec;
+
+public class KindImageLoader implements ImageLoader {
+
+  private Logger LOGGER = LoggerFactory.getLogger();
+
+  @Override
+  public void load(Project project, String image) {
+    LOGGER.info("Performing docker image loading with KiND!");
+    Exec.inProject(project).redirectingOutput(System.out).commands("kind", "load", "docker-image", image); 
+  }
+}
