+++ annotations/kind-annotations/src/main/java/io/dekorate/kind/config/DefaultKindConfigGenerator.java
@@ -17,7 +17,6 @@ package io.dekorate.kind.config;
 
 import io.dekorate.ConfigurationRegistry;
 import io.dekorate.config.DefaultConfiguration;
-import io.dekorate.kind.decorator.ApplyKindImageAutoloadConfiguration;
 
 public class DefaultKindConfigGenerator implements KindConfigGenerator {
