+++ annotations/kind-annotations/src/main/resources/META-INF/services/io.dekorate.ImageLoader
@@ -1,0 +1,1 @@
+io.dekorate.kind.KindImageLoader
