+++ annotations/servicebinding-annotations/src/test/java/io/dekorate/servicebinding/config/ServiceBindingConfigGeneratorTest.java
@@ -110,6 +110,6 @@ public class ServiceBindingConfigGeneratorTest {
     assertThat(tempDir.resolve("kubernetes.json")).doesNotExist();
     assertThat(tempDir.resolve("kubernetes.yml")).doesNotExist();
 
-    assertThat(result).hasSize(7);
+    assertThat(result).hasSize(6);
   }
 }
